[Android] LanguageSettings were modularized
The only reason we used BravePreferenceFragment is to have a close button in the
settings. This will be re-done in the context of this issue https://github.com/brave/brave-browser/issues/10375

Chromium change:
https://chromium.googlesource.com/chromium/src.git/+/f3c257e9298faf4da38daa674571c7ad08873106
[Android] Move LanguageSettings to //chrome/browser/language

This CL moves LangaugeSettings.java and the associated XML files out of
//chrome/android to the //chrome/browser/language module.

There are no UI changes from this CL.  A follow up CL will move the
language related tests to //chrome/browser/language.

Bug: 1124096