Adapt the way we extend ProfileManager's GetLastUsedProfileName()
This used to be a method of ProfileManager we could override from our
BraveProfileManager class, but now it has been moved as a private
helper function in an anonymous namespace in profile_manager.cc, so
we can't keep using the same approach.

Instead, just define a BRAVE_GET_LAST_USED_PROFILE_BASENAME via a
chromium_src override and inject it directly in the private function.

Chromium change:

https://chromium.googlesource.com/chromium/src.git/+/c404e341e3693efe7c71c91da55de8b54d3bde66

commit c404e341e3693efe7c71c91da55de8b54d3bde66
Author: Alex Ilin <alexilin@chromium.org>
Date:   Tue Apr 6 12:02:53 2021 +0000

    [profiles] Clean up ProfileManager::GetLastUsedProfileName()

    "ProfileName" is ambiguous because it can refer to both
    - the profile directory within the user data directory, and
    - the visible profile name.

    This CL renames GetLastUsedProfileName() to
    GetLastUsedProfileBaseName(). In addition, this method was only used
    internally, and this CL moves it to anonymous namespace in the .cc file.

    Bug: 1195201