Fixes GN check for recent_tabs_sub_menu_model_unittest.cc.
Error:

ERROR at //chrome/browser/ui/toolbar/recent_tabs_sub_menu_model_unittest.cc:44:11: Can't include this header from here.
          ^-----------------------------------
The target:
  //brave/browser/ui/toolbar:brave_app_menu_unit_test
is including a file from the target:
  //components/tab_groups:tab_groups

It's usually best to depend directly on the destination target.
In some cases, the destination target is considered a subcomponent
of an intermediate target. In this case, the intermediate target
should depend publicly on the destination to forward the ability
to include headers.

Dependency chain (there may also be others):
  //brave/browser/ui/toolbar:brave_app_menu_unit_test -->
  //chrome/browser/ui:ui --[private]-->
  //components/tab_groups:tab_groups